Add Paul Hsieh's SuperFastHash and test cases.
This is an implementation of the incremental version of the
SuperFastHash.  The main difference is that the hash function is seeded
with 0 instead of the length of the data being hashed (since we don't
know the full length of the data when we start.)

This routine is licensed under the LGPL to match the terms of the
original source code of which this is a translation.